Solution to Input Two Floats and Print Their Product

+++ Input Two Floats and Print Their Product@@6788d5012db25efec946f93e/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 int main(){
     flaot a,b;
-    int product;
+    float product;
     scanf("%f%f",&a,&b);
-    printf("Product: %d", product);
+    printf("Product: %.2f", product);
     return 0;
 }
